+++ evm/src/main/java/org/hyperledger/besu/evm/processor/AbstractMessageProcessor.java
@@ -80,6 +80,15 @@ public abstract class AbstractMessageProcessor {
    * @param evm the evm
    * @param forceDeleteAccountsWhenEmpty the force delete accounts when empty
    */
+  /**
+   * The EVM this processor runs frames on.
+   *
+   * @return the EVM
+   */
+  public EVM getEvm() {
+    return evm;
+  }
+
   AbstractMessageProcessor(final EVM evm, final Set<Address> forceDeleteAccountsWhenEmpty) {
     this.evm = evm;
     this.forceDeleteAccountsWhenEmpty = forceDeleteAccountsWhenEmpty;
@@ -181,6 +190,7 @@ public abstract class AbstractMessageProcessor {
     frame.getWorldUpdater().commit();
     traceFrameExit(frame, "SUCCESS");
     frame.getMessageFrameStack().removeFirst();
+    frame.returnStackToPool();
     frame.notifyCompletion();
   }
 
@@ -203,6 +213,7 @@ public abstract class AbstractMessageProcessor {
    */
   private void completedFailed(final MessageFrame frame) {
     frame.getMessageFrameStack().removeFirst();
+    frame.returnStackToPool();
     frame.notifyCompletion();
   }